Condensed API into one app to reduce cross-app referencing.
