--- server/buildings/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class BuildingsConfig(AppConfig):
-    name = 'buildings'
