+++ server/goodbullapi/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class GoodbullapiConfig(AppConfig):
+    name = 'goodbullapi'
