+++ server/goodbullapi/models/__init__.py
@@ -1,0 +1,5 @@
+from .courses import Course
+from .sections import Section, Meeting
+from .buildings import Building
+
+__all__ = ['Building', 'Course', 'Section', 'Meeting']
