+++ server/goodbullapi/models/buildings.py
@@ -12,3 +12,4 @@ class Building(models.Model):
 
     class Meta:
         ordering = ('abbr',)
+
