+++ server/goodbullapi/serializers/__init__.py
@@ -1,0 +1,4 @@
+from .courses import CourseSerializer
+from .sections import SectionSerializer
+from .buildings import BuildingSerializer
+__all__ = ['CourseSerializer', 'SectionSerializer', 'BuildingSerializer']
