+++ server/goodbullapi/views/__init__.py
@@ -1,0 +1,4 @@
+from .courses import CourseViewSet
+from .buildings import BuildingViewSet
+from .sections import SectionViewSet
+__all__ = ['CourseViewSet', 'BuildingViewSet', 'SectionViewSet']
