+++ server/server/urls.py
@@ -17,8 +17,6 @@ from django.contrib import admin
 from django.conf.urls import include, url
 
 urlpatterns = [
-    url(r'^buildings/', include('buildings.urls')),
-    url(r'^courses/', include('courses.urls')),
-    url(r'^sections/', include('sections.urls')),
-    url('admin/', admin.site.urls)
+    url('admin/', admin.site.urls),
+    url(r'^', include('goodbullapi.urls'))
 ]
